+++ middleman-core/lib/middleman-core/extensions/minify_css.rb
@@ -1,20 +1,15 @@
 # Minify CSS Extension
 class Middleman::Extensions::MinifyCss < ::Middleman::Extension
-  option :compressor, nil, 'Set the CSS compressor to use.'
   option :inline, false, 'Whether to minify CSS inline within HTML files'
   option :ignore, [], 'Patterns to avoid minifying'
-
-  def initialize(app, options_hash={}, &block)
-    super
-
-    app.config.define_setting :css_compressor, nil, 'Set the CSS compressor to use. Deprecated in favor of the :compressor option when activating :minify_css'
-  end
+  option :compressor, Proc.new {
+    require 'sass'
+    SassCompressor
+  }, 'Set the CSS compressor to use.'
 
   def after_configuration
-    chosen_compressor = app.config[:css_compressor] || options[:compressor] || SassCompressor
-
     # Setup Rack middleware to minify CSS
-    app.use Rack, :compressor => chosen_compressor,
+    app.use Rack, :compressor => options[:compressor],
                   :ignore     => Array(options[:ignore]) + [/\.min\./],
                   :inline     => options[:inline]
   end
@@ -36,9 +31,12 @@ class Middleman::Extensions::MinifyCss < ::Middleman::Extension
     # @param [Hash] options
     def initialize(app, options={})
       @app = app
-      @compressor = options[:compressor]
-      @ignore = options[:ignore]
-      @inline = options[:inline]
+      @ignore = options.fetch(:ignore)
+      @inline = options.fetch(:inline)
+
+      @compressor = options.fetch(:compressor)
+      @compressor = @compressor.to_proc if @compressor.respond_to? :to_proc
+      @compressor = @compressor.call if @compressor.is_a? Proc
     end
 
     # Rack interface